utils: security: fix security issues raised by gosec
Gosec complained about possible decompressing bombs when copying
files from remote locations or extracting files from remote locations.

I limited the size of files that cpackget can download and the file
that can be decompressed from a zip file to 20G.

Also, when inflating decompressed files, there is a check for "../" in
the file being decompressed.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>