signal: trap user termination requests
This patch introduces a monitoring thread that traps
Ctrl+C.

The patch consists of detecting user termination requests
between time consuming tasks in cpackget. For now there's 2:

1. Downloading packs
2. Extracting packs

If the user requests early termination on other tasks, there
will not be any effect, until some of the operations above
is executed.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>