installer: normalize paths
By default, CMSIS_PACK_ROOT environment variable is defined in a Windows
environment as C:\SOME_PATH. Both bash and cmd work fine with that approach.

In git-bash/mingw, if the user tries to "export CMSIS_PACK_ROOT=C:\SOME_PATH"
or "--pack-root/-R C:\SOME_PATH", bash automatically removes the backwards
slash "\" because it is considered to be a escape character, so it becomes
"C:SOME_PATH", which is a valid file name. It's wrong, but valid.

I wrongly assumed that this would throw an error. The colon (":") character
is not allowed in directory names on Windows (https://docs.microsoft.com/en-us/windows/win32/fileio/naming-a-file?redirectedfrom=MSDN#naming-conventions).
But in really, that does not fail in neither bash nor cmd:
* in bash, this becomes a valid name "C:SOME_PATH" which Windows later
  translates it to the unicode character "61498": https://wutils.com/unicode/61498
  which is valid.
* in cmd, this becomes "SOME_PATH" only, which is still valid.

To fix that, users should either escape all backwards slashes
"C:\\SOME_PATH\\SUB_DIR" or wrap it around single/double quotes
`cpackget -R "C:\SOME_PATH"`.

Another finding is that bash paths formatted as "/c/SOME_PATH" are always
translated to "C:/SOME_PATH" with the forward slash. And this is where
cpackget gets confused, because when it lists installed packs, it all start
as "C:\SOME_PATH". This means that cpackget will never receive paths in "/c/"
format. It seems this is just a visualization feature of bash.

By using the correct approach to set the path, cpackget works just fine.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>