make: Default to building for the host
If $OS is not set when running make, the Makefile currently defaults to
building for "linux", not the current host OS. Call out to uname to
determine the OS to use instead, so we can pick a default value for OS
based on the host system. This fixes auto-detection of the OS on macOS.

On Windows, %OS% is already set as an environment variable and we'll
skip calling uname, which is likely not available on Windows.

Fixes #51