installer: look up pack without version in index
If there is not a installed version of a pack, cpackget would never
figure the URL when installing a pack which version is not the
one in index.pidx.

Ex:

    $ cpackget -R .packs init https://keil.com/pack/index.pidx
    $ cpackget -R .packs pack add ARM.V2M_MPS3_SSE_300_BSP.1.2.0  # (note, at this time version 1.3.0 is the latest version of that pack)
    $ I: Using pack root: ".packs"
    $ I: Adding pack "ARM.V2M_MPS3_SSE_300_BSP.1.2.0"
    $ E: URL for the pack cannot be determined. Please consider updating the public index. Ex: cpackget --force index https://keil.com/pack/index.pidx

Except that version 1.2.0 is actually present in ARM.V2M_MPS3_SSE_300_BSP.pdsc.

The problem has now been fixed.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>