installer: pdsc: support removing local packs using PDSC file path
This allows cpackget to remove local packs, installed via local_repository.pidx,
by using a couple of forms of PDSC file path:

    $ cpackget rm Vendor.Pack.pdsc
    $ cpackget rm path/to/Vendor.Pack.pdsc

The first example looks for all PDSC tags in local_repository.pidx
with Vendor and Pack present.

The second example is a bit more restrictive. It only removes the
PDSC tags that contain Vendo, Pack and that the URL tag attribute
contains "path/to". The user might wish to specify the full path
of the PDSC file if wanted.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>