utils: packs: fix number of capturing groups in version regex pattern
Cpackget uses the regex for semantic versioning from PackIndexFile [1]
spec and I have figured that it is matching more than it should:

From the regex in PackIndexFile spec, this is the pattern snippet for
the "quality" part of the version (note that there is 3 capturing groups,
but the last matches many due to "*", causing it to generate a
**variable number of matches**):

     -                      alpha                   .0 (.5 is a separate match due to *)
    (-(0|[1-9][0-9]*|[0-9]*[a-zA-Z-][0-9a-zA-Z-]*)(\.(0|[1-9][0-9]*|[0-9]*[a-zA-Z-][0-9a-zA-Z-]*))*)?

Running it against "-alpha.0.5" will generate 4 matches:
"-", "alpha", ".0" and ".5".

Now take a look at the same regex snippet from Semver.org [2]:

       -                   alpha                     .0 (.5 matches due to *, except that now it belongs to the broader match "alpha.0.5")
    (?:-((?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?

The extra `?:` transforms the inner regex into a non-capturing group [3]
telling the engine not to include that in the matches, leaving the entire
snippet with only a single fixed capturing group. If we run the same regex on
"-alpha.0.5", we will get a single match "alpha.0.5".

Cpackget uses this regex to both match a given pack version and also to separate
it from the pack name. The latter was buggy because of the uncertain number of
matches that the regex could generate. After adding `?:` to reduce the number
of capturing groups, the bug was fixed.

[1] https://open-cmsis-pack.github.io/Open-CMSIS-Pack-Spec/main/html/packIndexFile.html
[2] https://semver.org/
[3] https://www.regular-expressions.info/brackets.html#noncap

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>