installer: test: fix intermittent tests on Windows
Some tests might fail on Windows intermittently due to how file
mod times are handled there. There was an earlier attempt to fix
this issue by sleeping for a couple of seconds before checking
for time mod differences. While that worked for a while, it started
popping up again.

So now, instead of sleeping only on Windows runtimes, it brings
back the original mod time of pack.idx so that if the file gets
touched, delta time will be huge enough.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>