installer: pack: don't install any pack version that is not available (#48)
The following patch rejects installing any pack version with a modifier that is not possible to satisfy. It won't even try to fetch them.

Both public and local indexes will be searched, and the error message provides info about the first possible version to install.

Two new sanity checks are provided to test this behaviour.