Add cryptography module and integrity checking features
This patch adds a new cryptography module, and two new commands to verify pack integrity.

'cpackget checksum-create' takes a local .pack and creates a .checksum file, similar to the typical digest files (like .sha256). It contains a list of all the pack files and their digest, whose cryptographic hash function is provided in the filename ('TheVendor.ThePack.1.2.3.sha256.checksum', for example).

'cpackget-verify' takes that .checksum file and verifies the integrity of a .pack, assuming it was generated with the other command.