Make CMSIS_PACK_ROOT Read Only
This patch is quite big! It makes all files and directories
under $CMSIS_PACK_ROOT Read-Only! The ONLY one exception for
this is the .Local/local_repository.pidx file, since it will
be written over and over when other tools add more local packs
into the repository.

I chose to do it all in a single commit because this might
cause some trouble with users not used to have all their files
in read-only mode all the time. So if needed, this can be reverted
in one go.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>