http-improvements: add timeout flag for downloads
Adds a global `--timeout/-T` flag to set a maximum duration of HTTP downloads, in seconds.
This applies to every HTTP GET performed by cpackget. It's set to 0/disabled by default as
to not make any assumptions on the network quality out-of-the-box.

However, if using the new batch PDSC download features, it might speed up the process a lot.
Tests with `-T 7` performed in Windows and Linux show that it speeds the batch downloads,
ignoring stuck downloads and only accidentally "dropping-off" some PDSC files.

Closes #110.

Signed-off-by: Luís Tonicha <luis.tonicha@linaro.org>