commands: crypto: generate/verify .checksums
Adds two new commands, "signature-create"
and "signature-verify", which respectively create
a detached PGP signature of a .checksum file, and
verifies it against its counterpart.

A private GPG key can either be generated by the "-create"
command, or passed through a flag. This key needs to
be in a PEM format, and also the signature.
Hence, it's recommended for both the key and the signature
to be created and verified using cpackget.

This is strictly for local usage, and currently
not using/storing keys in the OS keyring.

Signed-off-by: Luís Tonicha <luis.tonicha@linaro.org>