cryptography: create x509 full signatures
Adds a new command, signature-create-x509, which signs
a pack by setting it's zip comment field, to a specific signature scheme.
The signature can be composed of a x509 certificate (containing a vendor's
public key), the signed digest of the pack (digest sum of its contents), a PGP
signature or a mix of the first two.

This is the start of a refactor of the Signed Packs feature,
simplifying its implementation and guaranteeing non-repudiation
and authencity via the usage of Public Key Infrastructure (PKI).

Through the use of x509 certificates and/or PGP, vendors will be able
to publish & sign their packs with different degrees of complexity
and security assurance.