Signed packs rework and extension (#129)
* cryptography: rename PGP specific commands

* cryptography: create x509 full signatures

Adds a new command, signature-create-x509, which signs
a pack by setting it's zip comment field, to a specific signature scheme.
The signature can be composed of a x509 certificate (containing a vendor's
public key), the signed digest of the pack (digest sum of its contents), a PGP
signature or a mix of the first two.

This is the start of a refactor of the Signed Packs feature,
simplifying its implementation and guaranteeing non-repudiation
and authencity via the usage of Public Key Infrastructure (PKI).

Through the use of x509 certificates and/or PGP, vendors will be able
to publish & sign their packs with different degrees of complexity
and security assurance.

* cryptography: add full signature verification

Adds a signature-verify-x509 command that
inspects and validates a pack's signature "full"
signature scheme - which includes it's signed digest
and the x509 used to sign it.

* cryptography: add cert-only creation, verification and refactor

Adds signature creation and verification for "cert-only" mode,
which simply embeds an existing certificate into a pack.

Refactored "old" signature commands to reflect new scheme.

* cryptography: add PGP signing/verification

Adds the PGP signing scheme. Same logic
as the "full" mode, the user needs to provide the desired
PGP armored private key, which will create a base64
encoded PGP signature embed in the .pack.

In order to verify it, the public GPG key must be
present locally and provided.

Signed-off-by: Luís Tonicha <luis.tonicha@linaro.org>

* cryptography: cleanup and cmd testing

Cleanup and small refactor of the module.

Removes old PGP signature creation scheme as not
to confuse the user.

Adds testing for the signature-* commands.

Reverted stripLeadingZeros as cpackget will always
have a version without leading zeros.

Signed-off-by: Luís Tonicha <luis.tonicha@linaro.org>

* makefile: set coverage to atleast 85%

OS specific code is being introduced
which is not covered at all by the coverage
mechanism, and some error checking is not possible
to cover. Setting to 85% still maintains a high standard
with some margin for these "bad" cases.