installer: don't fetch PDSC from locally sourced packs
Disables trying to fetch the PDSC of a public pack, that
is provided by a local source (installed from the filesystem).
The pack gets installed normally as a public pack, but the
corresponding .pdsc won't be added to neither .Local or .Web
(since it wasn't downloaded).

This also fixes installing a public pack from a local source
when no internet connection is found.

Closes #134.

Signed-off-by: Luís Tonicha <luis.tonicha@linaro.org>